Annotated maps per potential fire alerts
Generate a map per potential fire alert that contains a highlighted pie
slice showing the potential area for the fire.  The map file is placed in
the same location as annotated images and videos, and link published in
both SQL DB and pubsub.
Bonus factorization of code to get central heading per camera